Update user controller

+++ backend/controllers/userController.js
@@ -1,11 +1,28 @@
 const User = require('../models/user');
+const { checkMissingFields } = require('../utils/validation');
+const mongoose = require("mongoose");
 
 const createUser = async (req, res) => {
   try {
-    const user = await User.create(req.body);
+    fields = req.body || {};
+
+    const requiredFields = {
+      email: "Email",
+      firstName: "First Name",
+      lastName: "Last Name",
+      hashedPassword: "Password",
+      age: "Age",
+      gender: "Gender",
+      weight: "Weight",
+      height: "Height"
+    };
+
+    if (checkMissingFields(res, fields, requiredFields)) return;
+
+    const user = await User.create(fields);
     res.status(201).json(user);
   } catch (err) {
-    res.status(400).json({ error: err.message });
+    res.status(500).json({ error: err.message });
   }
 };
 
@@ -20,7 +37,12 @@ const getUsers = async (req, res) => {
 
 const getUserById = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ error: 'Invalid user ID format' });
+    }
+
     const user = await User.findById(req.params.id).populate('currentDietPlan');
+
     if (!user) return res.status(404).json({ error: 'User not found' });
     res.json(user);
   } catch (err) {
@@ -30,16 +52,24 @@ const getUserById = async (req, res) => {
 
 const updateUser = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ error: 'Invalid user ID format' });
+    }
+
     const user = await User.findByIdAndUpdate(req.params.id, req.body, { new: true });
     if (!user) return res.status(404).json({ error: 'User not found' });
     res.json(user);
   } catch (err) {
-    res.status(400).json({ error: err.message });
+    res.status(500).json({ error: err.message });
   }
 };
 
 const deleteUser = async (req, res) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({ error: 'Invalid user ID format' });
+    }
+
     const user = await User.findByIdAndDelete(req.params.id);
     if (!user) return res.status(404).json({ error: 'User not found' });
     res.json({ message: 'User deleted' });
